urlをgrouppage?groupname=からgrouppage_xxxxxに変更　新規ステータス定義できるようにした。ステータス変更はまだです
