+++ the_project/grouppage/templates/grouppage/status_new.html
@@ -1,0 +1,33 @@
+{% load static %}
+<! dtype html >
+<html lang="ja">
+    <head>
+        <meta charset="utf-8">
+        <title>{{title}}</title>
+    </head>
+    <body>
+    <h1>新しいステータス作成ページ</h1>
+    ステータスはプロジェクトで統一してある<br>
+
+    既存ステータス<br>
+    {% for status_detail in status_details %}
+    {{status_detail.detail}}<br>
+    {% endfor %}
+
+    <form action="/mainpage/project_{{projectid}}/grouppage_{{group.uuid}}/status_new/" method="post">
+        {% csrf_token %}
+    {{form}}
+    <input type="submit" value="登録">
+
+    
+    <br>
+    <a href="/mainpage/project_{{projectid}}/grouppage_{{group.uuid}}">{{group.group_name}}</a><br>
+
+    <a href="">グループ設定</a>
+    <a href="/mainpage/project_{{projectid}}">プロジェクトページに戻る</a>
+
+
+    </body>
+</html>
+
+
