+++ the_project/mainpage/grouppage.py
@@ -1,0 +1,5 @@
+from ..grouppage.view.roompage import *
+from ..grouppage.view.status_new import *
+
+#roompage→部屋ページ
+#status_new→新しいステータス作るとき
